Don't allow the same argument to be set twice in one call
The original Tcl implementation of named parameters supported specifying the same argument multiple times.  For example, "foo -b 1 -b 2" and it just used the last value set.

The C implementation was made much simplier, touching less code in fewer places, by making the forementioned practice an error.  We agreed to do that, changed one of the np tests accordingly, and the C implementation was developed and shown to pass all of the tests of the Tcl version.

However, we didn't bring the Tcl version into compliance with the C behavior.  This PR bring the Tcl version into compliance and so now once again the Tcl version will be able to pass all the tetss.