Update np.tcl
Adding an additional type of named_parameters. While the regular named parameters (e.g. -parameter) are defined with the procedure definition e.g.

np::proc test {{a 1} {b 2} -- {c 3} } { 
    return [info vars *]
}

the new ad-hoc named parameters are defined at runtime

test -a 5                          will have      a=5,    b=2,    c=3
test -a 5 =d 4                  will have      a=5,    b=2,    c=3,    d=4

of course any other suitable character could be chosen instead of '=' to identify these ad-hoc named characters, but the '=' is kind of charming.

Since this would be a new feature, the version number was bumped to 1.1.0.